refactor(foreman): promote condition helpers to a shared file
setCondition and hasCondition were duplicated in agentictask_controller.go
and fleetnode_controller.go respectively. A third writer (the AgentRelease
controller, PR3) would require a third copy, so promote them now.

Extract both into internal/foreman/controller/conditions.go (same package,
no import change for callers). Delete the local copies from both source
controllers. No behavioral change; existing tests pass unchanged.
